fiks resttemplate config
